refactor: Move functions into SharesOverviewScreen.dart
